+++ RobotsOnMars/Console/Parser.cpp
@@ -1,4 +1,11 @@
 #include "Parser.h"
+#include "../GameModes/ManualMode.h"
+#include "../Commands/Scan.h"
+#include "../Commands/ChangeMode.h"
+#include "../Commands/Grab.h"
+#include "../Commands/Move.h"
+#include "../Commands/Quit.h"
+#include "../Application/Manager.h"
 
 Parser::Parser() : commandIdx{0}, mapWidth{15}, mapHeight{15} {}
 
@@ -12,11 +19,11 @@ ICommand *Parser::GetCurrentCommand(Manager *manager, bool necessary)
 ICommand *Parser::ParseCommand(Manager *manager, bool necessary)
 {
     ICommand *currentCommand = nullptr;
-    size_t activeRobotsNum = manager->GetActiveRobots()->size();
+    int activeRobotsNum = manager->GetActiveRobots()->size();
     Robot *robot = manager->GetActiveRobots()->at(0).second;
     ModeBasement *currentMode = manager->GetActiveRobots()->at(0).first;
     string word = "";
-    size_t stepsNum = 0;
+    int stepsNum = 0;
 
     if (necessary)
     {
